ATK: FIX for requirement to include MathLib Quark for sc3-plugins
Contribution by @mtmccrea to resolve issue discussed:

https://github.com/supercollider/sc3-plugins/pull/124
https://github.com/supercollider/sc3-plugins/pull/125
http://new-supercollider-mailing-lists-forums-use-these.2681727.n2.nabbl
e.com/Fwd-sc-users-Re-SC3-plugins-missed-class-td7628763.html

At the moment, it appears the long term solution will be to move all
the language side elements of the ATK to a separate Quark. (The ATK
team has been considering this option.) In the short term, this current
solution removes the dependency on users installing MathLib when
installing sc3-plugins.